+++ src/write/adios_flexpath.c
@@ -982,7 +982,7 @@ set_format(struct adios_group_struct *t,
 // copies buffer zeroing out pointers to arrays 
 void *copy_buffer_without_array(void *buffer, FlexpathWriteFileData *fileData)
 {
-    ADIOST_CALLBACK_ENTER(adiost_event_fp_copy_buffer, fileData);
+    ADIOST_CALLBACK_ENTER(adiost_event_fp_copy_buffer, (int64_t)fileData);
     char *temp = (char *)malloc(fileData->fm->size);
     memcpy(temp, buffer, fileData->fm->size);
     FMField *f = fileData->fm->format[0].field_list;
@@ -993,7 +993,7 @@ void *copy_buffer_without_array(void *buffer, FlexpathWriteFileData *fileData)
         }
         f++;
     }
-    ADIOST_CALLBACK_EXIT(adiost_event_fp_copy_buffer, fileData);
+    ADIOST_CALLBACK_EXIT(adiost_event_fp_copy_buffer, (int64_t)fileData);
     return temp;
 }
